project/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
numpy==1.26.4
xgboost==2.1.1
scikit-learn==1.5.1
langchain==0.2.16
langchain-openai==0.1.23
faiss-cpu==1.8.0
pymupdf==1.24.9
python-dotenv==1.0.1
pydantic-settings==2.5.2
httpx==0.27.2